+++ src/hed/mcc/tls/MCCTLS.h
@@ -14,7 +14,7 @@ namespace Arc {
   public:
     MCC_TLS(Arc::Config *cfg);
   protected:
-    bool tls_random_seed(std::string filename, size_t n);
+    bool tls_random_seed(std::string filename, long n);
     bool tls_load_certificate(SSL_CTX* sslctx,
 			      const std::string& cert_file,
 			      const std::string& key_file,
